add instructions for running against minishift
Update the README with instructions for setting up the controller in a
dev environment using minishift.

Update the role binding YAML file to allow a developer to run the
controller from outside the cluster. We will likely want to do
something different when we get to really packaging this for
deployment.

Signed-off-by: Doug Hellmann <doug@doughellmann.com>