update operational status handling
Add a toggle so the user can control whether they want the machine
online or offline.

Move the status field to a label so it can be used when searching for
available hosts.

Simplify the error message field in the Status struct since we no
longer need the operational status value to be stored there.

Add tests showing that toggling the online value has the desired
effect.

Use some const values instead of literal strings scattered throughout
the code.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>