rework Reconcile() to do one task at a time
Rework the status management so we save the Status block at the point
where it is modified and so Reconcile() can stop after it does one
task. This avoids having to try to reload the BareMetalHost object to
get the latest value, which doesn't always work because of race
conditions. Requeuing the object for reconciliation ensures that the
next time Reconcile() runs it has the latest data available.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>