inline credentials secret loading to make error handling easier
The logic for loading a secret is simple enough now that having it in
a separate function actually complicates error handling for the case
where we do not have a secret name. Move the lookup inline in
Reconcile() so that case can be handled properly by setting the error
condition and returning without an error, hence waiting for the CR to
be updated before reconciling again.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>