Unvendor apimachinery
The types provided by k8s.io/apimachinery must work as a shared
dependency between servers and clients.  As such, the package cannot be
vendor.  When vendored, new instanced of `BareMetalHost` and similar
cannot be properly constructed.

For example, the `BareMetalStatus` struct has a meta field `LastUpdated`
which is of type `*k8s.io/apimachinery/pkg/apis/meta/v1/Time`.  The
import path for the `Time` struct is different/wrong when the meta
package is vendored in this repo.