simplify provisioner factory interface
With most of the provisioning data coming from the host object, we no
longer need to use a separate factory object and can just use a
function interface.

This change also moves the ironic URL constant into the ironic
module. If we need that to be configurable in the future, we will need
to change the factory interface to pass some configuration source in.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>