clean up operational status handling
Move the operational status value out of a label into the status block
of the host. This avoids cases where we have to update the host
metadata and status blocks separately, which can lead to odd behavior.

Also define a type for the status values and remove some of the unused
values. The list will need to be expanded separately.