move hardware profile out of a label
Store the hardware profile name in the status block instead of a
label. The actuator is not going to use the label to search through
the inventory anyway, and this lets us remove a potential race
condition of trying to update the metadata and status for the host at
the same time.