let the baremetal operator download the checksum from a url
The unit tests for the actuator are failing because we require it to
provide a full checksum for the image being provisioned. This will
allow it to just pass us the URL, and have the operator download it to
learn the checksum from there. Eventually we can just pass the URL to
ironic and let it do the download, but we have to upgrade to a newer
ironic for that.