clean up error handling in ironic provisioner
Remove the HasError() checks because that case was something that came
up in an earlier dev version and the Reconcile() checks should prevent
repeated calls on error now.

Add checks to see if the node we try to load is nil, which can come up
if ironic gets out of sync with the CRD. Fixing that out of sync issue
is a separate change.