improve host error handling
This change improves the way we handle errors detected each time
Reconcile() is invoked.

It starts by clearing the error message at the outset, to ensure that
any old error is removed if the situation is corrected and to help us
know when a new error is detected.  When we clear the error message we
only want to clear the operational status if it is set to error. This
allows Reconcile() to correctly track the state of the host.

Then after each significant step, in addition to checking for an error
condition we can retry (such as an error response from an API call) we
check if the host has an error set. In those cases we assume we cannot
retry, and simply stop reconciling until the host is modified again
later.

We remove error checking from the provisioner, leaving it up to
Reconcile() to avoid calling the provisioner when the host has an
error.