combine all delete handling
Without this change, hosts in an error state cannot always be deleted.
Rather than trying to handle each delete case separately, merge them
all into 1 function that explicitly ignores certain error cases to
build enough objects to tell the provisioner that we need to delete
the host and then allows the k8s API to do that when the host is
cleaned up.

Also clean up access of "ObjectMeta" properties of the host.