rely on ironic state to determine the next action
Move all provision state management out of the provisioner into
Reconcile() so it is consistent across drivers.

Determine the next driver action and store that status value before
taking any action so the state is consistent with what is happening.

Update the provisioner to look at the status of the node in ironic to
determine its next step within each API method. (This may not cover
all necessary states, but it covers the ones that were already
handled.)