improve error handling and retry provisioning
Add states for provisioning and power management errors.

Use a field in the provisioningResult structure to track whether the
current operation has produced an error. If so, update the error
status at the end of each action method. If not, clear the error
status at the end of each action method.

In the ironic provisioner, when we have a deployment failure wait for
ironic to provide a description of the error.

In the ironic provisioner, retry provisioning when the user updates
the image URL or checksum.

When the power status is changed, update the status field to reflect
the new status.