Add script to run a local Ironic for development.
Ironic should eventually run in the baremetal-operator pod.  Right now
our sample pod yaml includes a sidecar proxy container that redirects
connections to the expected local Ironic to an Ironic running
elsewhere.

This script runs a set of containers for Ironic built from metalkube
repos.  It's useful for development environments until we move these
containers into the cluster properly.

This is copied from a repo of scripts where a group is workgin on
metalkube integration with OpenShift, but with the OpenShift specifics
removed and made more generally useful.

The original script was here:
https://github.com/openshift-metalkube/dev-scripts/blob/master/04_setup_ironic.sh