let externally provisioned hosts be registered with ironic
The master nodes will have already been registered, but worker nodes
will not, so we need to let them pass through the registration phase.

This change updates the state diagram as well as Reconcile(). It also
adds an event for hosts leaving the "registering" state and about to
enter the "externally provisioned" state.