do not deprovision externally managed systems on delete
When we delete a host associated with an externally provisioned
system (i.e., a master), we do not want to deprovision it or power it
off because that can break the cluster. To support that, we split the
delete operation out of the deprovision method of the provisioner, so
the reconciler can call them appropriately.

The ironic provisioner delete logic is improved to force hosts into a
state where they can be deleted from ironic without side-effects.