make registration error a steady state
If a host fails registration with the same error message a second
time, that means we have reconciled it twice with the same data and it
still does not work. Rather than keep trying, which toggles the host
between "registering" and "registration error" indefinitely, leave the
host in the "registration error" state.