rename git repository
This change is part of renaming metalkube to metal3. We have renamed
the git repository, so in order to build this repo in its new location
it needs to have the imports updated. We do not want the full rename
in this branch, though, to minimize changes in a "stable" branch.