Update main.go
The current make-bm-worker tool does not include a boot MAC Address for your baremetal nodes which is a requirement for the provisioning.

Update templates.go

Updating the templates file to include in the yaml file the bootMACAddress

Update main.go

Changed flag from -bootmacaddress to -boot-mac

Update main.go

removed bootMACAddress from being a requirement. The boot-mac flag is optional.