Report RAM size in MiB
RAM doesn't always come in a round number of GibiBytes due to e.g.
reservations for video memory. It usually comes in a round number of
MebiBytes (although /proc/meminfo reports the size in kibiBytes), so use
that as the unit for reporting RAM size. This also happens to match the
units used by ironic-inspector.

Signed-off-by: Zane Bitter <zbitter@redhat.com>