Improve Storage HardwareDetails
- Report disk size in Bytes
- Don't assume storage type

Actual disk sizes vary greatly from round numbers of GibiBytes as a
result of different practices in labelling disks. The only common
denominator is Bytes, so report the size in Bytes and rename the field
to SizeBytes. This also happens to be the format in which
ironic-inspector reports disk sizes.

We only know from ironic-inspector whether a disk is rotational storage
(i.e. an HDD) or not. We can't know whether it is a SATA-attached SSD,
NVMe attached SSD, or something else. Ironic-inspector gets this
information directly from udev, so this is likely to be universally as
good as it gets. Just report the information we have available, and
rename the field to Rotational.

Signed-off-by: Zane Bitter <zbitter@redhat.com>