clear error when credentials are recognized as good
This change fixes a bug in which an existing host could be updated to
have bad credentials and then that mistake could be corrected but the
host status would not be fixed. Always clear any existing error
messages when the credentials are recognized as changed and good.

Also adds a unit test for that case.

Fixes #195