Add an idempotent reboot request
Allow a client to request that the server be rebooted by specifying the
current time as the RebootAfter time. Multiple clients can request this
(or not) without stepping on each other's toes, and without affecting
the target power state (so an attempted reboot can't power up a host
that another client has powered off).

Signed-off-by: Zane Bitter <zbitter@redhat.com>