Wait for power status to change
When attempting to power a host on or off with Ironic, we were declaring
success (updating the power status and dropping back to polling only
every 60s) as soon as we saw the target power state change in Ironic
even if the host had not yet reached that state, and also in the case
where we were unable to set the target power state because the host was
locked.

Ensure that the power state has actually changed before we consider the
change complete.

Signed-off-by: Zane Bitter <zbitter@redhat.com>