handle deprovisioning in two phases
The cluster-api-provider-baremetal implementation of deprovisioning
first removes the image reference and then removes the consumer
reference to ensure that a host is not grabbed and used while it is
being deprovisioned. This change fixes the state management to
understand that case and adds a test to demonstrate that the
reconciler logic works as expected.

Fixes #238