only skip inspection for fully provisioned hosts
Using the presence of a consumerRef value to indicate that a host is
in use means we have a race condition when a host is first registered
if the machine controller grabs the host and allocates it to a
machine before the host has finished inspection. To avoid the race,
use the WasProvisioned() function to test whether a host has actually
been provisioned instead of looking only at the consumer reference.

This update requires changing the expected behavior in an existing
unit test (because it was wrong) and adding a new unit test for the
new case.