update the BMC credentials given to ironic when they change
When we get new credentials for a server that has already been
registered with bad credentials we want to update the settings in
ironic and try again. This requires updating the driver_info data and
then attempting to manage the host again.

The provisioner method ValidateManagementAccess() needs to know if the
credentials have changed, but we cannot determine that by asking
ironic because it will not expose credentials through its
API. Therefore we have to keep track ourselves by adding a
TriedCredentials field to the host status struct. The controller
manages this field and passes a flag to the provisioner when the
credentials have been changed.

Clean up some of the rest of the error handling in the ironic
provisioner version of ValidateManagementAccess() to make it a bit
clearer.

Add a Match() method to the CredentialsStatus type for comparison
against the Secret directly and remove the CredentialsNeedValidation()
method.