update host status to track operation times
In order to gather metrics like how long it takes to perform a given
operation we have to track the start times. We might as well track the
end time as well as a way for the user to review the history of the
host (which is admittedly more interesting for provisioning and
deprovisioning operations than for others but it's nice to have the
data structure consistent).

This change adds fields to the structure to hold times we are
interested in.