handle adoption of provisioned and ready hosts
Rename actionManageExternallyProvisioned() to
actionManageSteadyState() and pass all hosts that are in a steady
state to it. This ensures that we adopt (reregister) any hosts that
are lost from ironic after they have been provisioned or that are
"ready" and available for provisioning, along with the previous work
to handle externally provisioned hosts.