Use leader election mechanis from controller-runtime library
controller-runtime library provides more correct way to set leader among pods,
it is done via config map annotation, that includes leader lease.
In case when the pod fails to update the lease under the config map, the other pod
will aquire the leader.

Otherwise the operator-sdk do not provide the lease feature and just creates the
config map, so until the config map exists, other pod can not aquire the leader,
but the config map will be exist as long as pod exists(also in terminating state),
so it will prevent HA in cases when node not responsive from some reason(network, kernel panic, ...).

Fixes #295