Remove *Error states
Changing the provisioning status is not how we actually indicate errors:
we have the operational status for that. The existence of these states
also has the effect of obscuring which state the Host was in (and hence
what we were trying to do) when the error occurred. For example, a
registration error can occur during provisioning.

Remove the RegistrationError, ProvisioningError, and
PowerManagementError states, in line with what the state machine diagram
envisions, but also remove the Error states from the diagram as they are
now orthogonal to the provisioning state.