Handle conflicts consistently in ValidateManagementAccess()
When we get an HTTP 409 response in ValidateManagementAccess, we should
retry after a delay. In one case we were not delaying (i.e. retrying
immediately), and in another we added a delay before moving to the next
state but reported that the management access was valid when that had
not been verified.

Handle both with a retry after a delay.

Since this was the only case where the provisioner would request a delay
while reporting success, this also removes the code handling that case
in actionRegistering(). Once the host is registered we always want to
requeue immediately so we can move to the next state.