Set error message on inspection error
When an error occurred during inspection, we weren't setting the Requeue
flag, which is required for the CR to be updated (although it doesn't
actually cause the reconcile to be requeued, because of the error).

This also meant that an error event would be published each time the CR
was reconciled, even if the error had not changed. The error event also
wrongly stated that it was a registration error.

Finally, we weren't actually changing the Host's ProvisionState to
RegistrationError (since changes weren't written), so don't start now.