Return actionResult from ReconcileState()
This enables us to simplify the logic in the top-level Reconcile()
handler. We can now use the Dirty() method of the result to determine
whether we need to write the status, rather than overloading the Requeue
field. To determine whether the action failed, we can look for a result
of type actionFailed instead of looking at the Host status (which
doesn't tell us if the failure was pre-existing). Finally, the
RequeueDespiteError field is no longer needed as we always return
actionComplete (not actionFailed) when the provisioning state changes to
something other than an error state.