Handle registration prior to state machine
The registration state of the Host is effectively orthogonal to the
provisioning state so, instead of bouncing back to the Registering state
whenever the credentials change, ensure the Host is registered with the
current credentials prior to handling the current state.

The Registering state remains as a transition state to the rest of the
state machine - the Host will not exit the Registering state until it
has been registered at least once.