Transition to Deprovisioning state when deleting
If a Provisioned or Provisioning host was deleted, we used to commence a
delete and handle deprovisioning in the course of that. Instead, transition
to the Deprovisioning state for deprovisioning before proceeding to
deletion. This ensures that deprovisioning has only one implementation in
the controller.

Since deprovisioning may (in theory) fail and we (presumably) don't want
to make the Host undeletable, we continue to skip straight to deleting
if we get a failure in the deprovisioning state. However, note that in
practice the provisioner never returns an ErrorMessage from
Deprovision() (i.e. it never gives up trying to deprovision).