Add a finite state machine implementation for BareMetalHost
This is currently unused.

Once we switch over to this implementation, the next action of the Host
will depend not only on the difference between the spec and the status, but
also on the current state.

This means that e.g. a deprovisioning operation will run to completion even
if the user restores the image details to the spec. Currently the operation
will complete in Ironic but the Host would return to the Provisioned state.

The allowable state transitions are based on those in the state diagram
design doc, with some modifications to prevent substantive changes from
existing behaviour. (For example, the Host can transition to the
Registering state from any other state, and then transition back again once
registration is complete.)