Clean up logging of Reconcile() result
Since we separated the meanings of Dirty() and Result().Requeue in
actionResult, we can simply look at the Requeue value to determine if
reconciliation has failed (except in the case where deletion of the
resource has been completed). There's no need to base this on the result
type. Split this implementation into a separate function for clarity.