Delay computing stats until the op is recorded
If there is a conflict and we fail to publish the status update to the
host CR then we will count the operational metrics multiple
times. This change uses a callback list to update the counters only
after saving the host, at which point we know that the operation is
successful.