Fix crash when cancelling provisioning
If the operator has started provisioning a Host, and the user then
cancels the provisioning (by removing the Image from the Spec), the Host
should transition to the deprovisioning state immediately and not wait
for provisioning to complete.

This was not occurring, because NeedsDeprovisioning() assumed that the
Host had to be fully provisioned before it would need to be
deprovisioned, which is not the case. If the Image struct was set to nil
(rather than just empty) then this caused a crash because the Ironic
provisioner would attempt to access the structure unconditionally inside
Provision().

Since the state machine calls NeedsDeprovisioning() only from the
Provisioning and Provisioned states, we can now eliminate the problem by
always returning true from NeedsDeprovisioning() whenever there is no
image to be provisioned.

Fixes #377
Fixes #342

Signed-off-by: Zane Bitter <zbitter@redhat.com>