Don't check ExternallyProvisioned flag
The state machine defines that the Host transitions to the
ExternallyProvisioned state only from the Ready state; and when leaving
the ExternallyProvisioned state always transitions back to the Ready
state. However, the user can modify the Spec.ExternallyProvisioned flag
at any time. It is unsafe to take this flag into account e.g. in the
middle of a provisioning operation, as it can leave the Host in an
undefined state that is not recorded in the CR.

Signed-off-by: Zane Bitter <zbitter@redhat.com>