Avoid starting inspection twice
If we end up reconciling the Host again after starting an inspection in
Ironic, but before the IntrospectionStatus is available in
ironic-inspector, we don't want to attempt to start another inspection.
Check if the host is already in the correct state in Ironic before
trying to change the state to 'inspecting'.

This issue was identified in #393, but may not be the cause.