Add idrac-redfish url type
The idrac BMCs have partiular handling of boot mode as configuration
change requests which need to be honored only at power state changes
as opposed to immediately, because this can conflict and over-write
existing configuration jobs.

While we shouldn't be in this state without more advanced bmc
features being leveraged, it is best to use the vendor supported
hardware type and interfaces for using redfish with their hardware.

This boot interface defaults to using iPXE, as opposed to virtual
media. A separate idrac specific redfish virtual media interface
already existed in the baremetal-operator.