add root device hints to BareMetalHost
Add a new field spec.rootDeviceHints to the API with sub-fields for
different types of hints based on what Ironic supports today.

Add a new field status.rootDeviceHints to record the values used when
provisioning the host.

Change the Provisioner API to return the new public struct with the
more detailed device hints so that the values in the hard-coded
profiles can be stored directly in the new status field.

If the user does not provide explicit hints, use the values from the
profile.

Implements https://github.com/metal3-io/metal3-docs/blob/master/design/user-defined-root-device-hints.md

Co-Authored-By: Mika Koskimaki <mika.koskimaki@est.tech>
Co-Authored-By: Doug Hellmann <dhellmann@redhat.com>
Signed-off-by: Doug Hellmann <dhellmann@redhat.com>