Switch boot mode to pxe
In order to enable secure boot, we need to get
rid of ipxe and start using pxe, to be able to use
signed images from the distros. This commits changes
this mode, and in parallel the ironic-image used in
the deploy will be changing its configuration to use
pxe instead.

Signed-off-by: Yolanda Robla <yroblamo@redhat.com>
Fixes-bug: #1814164
Depends-on: https://github.com/metal3-io/ironic-image/pull/144