Don't assume IPv4 in introspection data
This can be one of two keys, depending on whether the host
got an ipv4 or ipv6 address, so use the ipv6 key in the case
where the ipv4 one is unset.

Currently in a single-stack ipv6 deployment the IP ends up
empty like:

Hostname:     worker-0.ostest.test.metalkube.org
    Nics:
      Ip:

But with this change Ip should be correctly populated.