make hack/gofmt.sh return error if it wants to make changes
gofmt does not return an error code if it can successfully find
changes, but for the prow linter job we want this script to return an
error if the code is not already formatted. Look for the list of files
gofmt would change, and if the list is not empty then show the diffs
it would make.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>