set up check to ensure operator-sdk generation is done
Add a make target 'generate-check' to ensure that both the code and
CRDs have been regenerated for any change.

Add hack/Dockerfile.operator-sdk to build an image containing the
operator-sdk and other dependencies for the generator check.

Add hack/generate.sh to run the check in a container using the image
built by the new Dockerfile.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>