Add support for mutual TLS and custom CA for ironic
The IRONIC_TRUSTED_CA_FILE variable can be used to give a custom CA that
will be trusted when connecting to ironic and inspector.

When the environment variables IRONIC_CERT_FILE, IRONIC_KEY_FILE and
IRONIC_TRUSTED_CA_FILE are all set, mutual TLS will be used to
communicate with the ironic endpoints.

The http.Client with the mTLS configuration will load the client
certificate from the disk every time a request is made. Therefore, it
allows to rotate the client certificate without a restart. The CA file
is only loaded when a new provisioner is created. This happens during
every reconciliation. Therefore, a rotation of the CA certificate
without restart is also supported.