Round clock frequency to integer when setting HW details
Since float is not considered a valid type we will eventually have to
convert this field to an integer. That's probably going to involve some
very not-fun contortions regardless (see metal3-io/metal3-docs#101), but
at the very least it's not helping anybody to continue creating Hosts
with fractional values that will be non-round-trippable later.