Rename "Discovered" state to "Unmanaged" in state diagram
This state isn't actually implemented in the code, so before we
implement it rename it to "Unmanaged". This reflects the fact that what
this state really means is that the Host has been created without
attempting to provide credentials. That may have transpired through a
discovery process or some other mechanism.