Always move to Registering state immediately
Bring the code into line with the state machine documentation, which
says that once the Host is reconciled it will immediately move into one
of the states (registering or unmanaged) and not remain with an empty
provisioning state. Since the Unmanaged state has not been implemented
yet, move unconditionally to Registering.

Previously, we waited until the the BMC credentials had at least once
been deemed worthy of trying (present, correctly formatted) before the
Host would move to the Registering state.