Implement the Unmanaged state
Newly-created Hosts enter the Unmanaged state if no BMC details are
provided in the resource Spec.

As soon as details are provided - regardless of whether they are
complete or correctly formatter - the Host will return to the
Registering state. Even if all of the details are later removed, the
Host will not return to the Unmanaged state.

This is in line with how this state is documented in the state machine
diagram, but it was never previously implemented.