Set RegistrationError if managed host is missing BMC data
When a host is discovered or otherwise created without any BMC data, it now
goes into an Unmanaged state, in which buildAndValidateBMCCredentials()
is not called.

Once a host leaves the unmanaged state (i.e. the BMC address and/or
credentials have been supplied at least once), any time the BMC address
and/or credentials are missing it is due to a configuration error, not
to having newly-discovered a host.

Since the unmanaged state handles newly-discovered hosts, in the error
handling for buildAndValidateBMCCredentials we can now flag an error if
any of the data is missing. This prevents transitions like a Host going
back into a "discovered" operational status despite having been
provisioned already before the credentials were removed. That should be,
and now is, treated as an error.