Don't write status annotation
The code to write a status annotation can get stuck in an infinite hot
loop, and the upstream fix to it seems highly prone to further issues.

We don't need (or want) to write a status annotation in OpenShift
(though we do need to be able to read one), so disable the offending
code.

https://bugzilla.redhat.com/show_bug.cgi?id=1851530