pkg/bmc: refactor to support all schemes
All of the redfish BMC's should support `+http` or `+https` explicitly,
however it's not configured for any of the virtualmedia BMC's. This
refactories the registerFactory code to allow specifying schemes for a
BMC, and ensuring the schemed versions exist for every one.

This keeps it explicit, so we can say which BMC's support specifying a
scheme.