add command line options and config for starting controller
The way a controller is set up under kubebuilder is different from
operator-sdk, so some of the option definitions have moved to the main
program and the provisioner factory becomes a public member of the
reconciler type.

The concurrency limit is still managed via an environment variable
passed as an option when the reconciler is set up on the manager.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>