change default hardware profile to not set root device hints
Change the default behavior to not set an explicit root device hint
based on the hardware profile. This relies on Ironic's default
behavior of choosing the first disk that is bigger than 4GB to hold
the image, which can often bypass issues such as recovery drives or
other devices appearing as `/dev/sda`.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>