Keep retrying failed operations with backoff
Improve the reconciliation loop whenever an action failure is detected
(or credential error) by applying a retry pattern with exponentional
backoff with jitter to avoid service overloading - currently affected
statuses are: Deprovisioning, Externally Provisioned, Inspecting,
Provisioned, Provisioning and Registering.
A new `ErrorCount` field has been added in the `BareMetalHostStatus`
to support such behavior: it gets incremented every time an action
failure is recorded, and it is cleared out when the action completes
successfully.