handle externally provisioned hosts without image settings
Update the IronicProvisioner to deal with externally provisioned hosts
that have no image settings by recording an error message of our
own instead of leaking the Ironic error.

When the host does have image settings, ensure they are sent to Ironic
before telling it to adopt the host, in case they were not present
when the host was registered (this is similar to what we do for
provisioning).

Addresses #608

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>