only pass ironic image settings when they are available
Now that getUpdateOptsForNode() is being called before inspection, we
may be using it when the image values are not set for the host. We
don't want to pass empty values to ironic, so the logic is
restructured to only set values associated with the image when there
are values to send.

The general structure of the function is reorganized at the same time
to deal with values coming from the hardware profile first, right
after the profile is fetched, and to handle all of the image values
next.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>